# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for User Service",
        "title": "API Documentation",
        "contact": {},
        "version": "0.2"
    },
    "paths": {
...
        "/numerical-method/linear-algrebra/matrix": {
            "post": {
                "description": "Create the matrix result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matrix"
                ],
                "summary": "Create Matrix Result",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqMatrix"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Matrix"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/linear-algrebra/matrix-iteration": {
            "post": {
                "description": "Create the matrix iteration result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matrix Iteration"
                ],
                "summary": "Create Matrix Iteration Result",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqMatrixIteration"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.MatrixIteration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/linear-algrebra/matrix-iteration/{id}": {
            "get": {
                "description": "Get the matrix iteration result by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matrix Iteration"
                ],
                "summary": "Get Matrix Iteration Result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matrix Iteration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Graphical"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/linear-algrebra/matrix/{id}": {
            "get": {
                "description": "Get the matrix result by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matrix"
                ],
                "summary": "Get Matrix Result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matrix ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Graphical"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/numerical-diff": {
            "post": {
                "description": "Create numerical diff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "numerical-diff"
                ],
                "summary": "Create numerical diff",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqNumericalDiff"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.NumericalDiff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/numerical-diff/{id}": {
            "get": {
                "description": "Get numerical diff by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "numerical-diff"
                ],
                "summary": "Get numerical diff",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Numerical diff id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.NumericalDiff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Graphical": {
            "type": "object",
            "properties": {
                "equation": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "scan": {
                    "type": "number"
                }
            }
        },
...
        "models.Matrix": {
            "type": "object",
            "properties": {
                "constant_data": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matrix_data": {
                    "type": "string"
                },
                "matrix_size": {
                    "type": "integer"
                }
            }
        },
        "models.MatrixIteration": {
            "type": "object",
            "properties": {
                "constant_data": {
                    "type": "string"
                },
                "e": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "matrix_data": {
                    "type": "string"
                },
                "matrix_size": {
                    "type": "integer"
                }
            }
        },
...
        "models.NumericalDiff": {
            "type": "object",
            "properties": {
                "function": {
                    "type": "string"
                },
                "h": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "order": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "validations.ReqMatrix": {
            "type": "object",
            "properties": {
                "constant_data": {
                    "type": "string"
                },
                "matrix_data": {
                    "type": "string"
                },
                "matrix_size": {
                    "type": "integer"
                }
            }
        },
        "validations.ReqMatrixIteration": {
            "type": "object",
            "properties": {
                "constant_data": {
                    "type": "string"
                },
                "error": {
                    "type": "number"
                },
                "matrix_data": {
                    "type": "string"
                },
                "matrix_size": {
                    "type": "integer"
                }
            }
        },
...
        "validations.ReqNumericalDiff": {
            "type": "object",
            "properties": {
                "function": {
                    "type": "string"
                },
                "h": {
                    "type": "number"
                },
                "order": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                }
            }
        },
...
    }
}